Move new functions into main util lib
