+++ Clausius.py
@@ -21,110 +21,10 @@ from __future__ import division, print_function, absolute_import
 import time
 import datetime as dt
 
-from ligmos.utils import packetizer
 from ligmos.workers import connSetup
 
 from ultimonitor import apitools as api
 from ultimonitor import confparser, printer
-
-
-def systemStats(printerip):
-    """
-    Query the printer for memory
-    """
-    endpoint = "/system/memory"
-    mem = api.queryChecker(printerip, endpoint)
-    if mem != {}:
-        mem.update({"free": mem['total'] - mem['used']})
-
-        pkt = packetizer.makeInfluxPacket(meas=['system'],
-                                          fields=mem,
-                                          tags=None)
-    else:
-        # Silly.
-        pkt = []
-
-    return pkt
-
-
-def tempStats(printerip, timeoffset=None):
-    """
-    Query the printer for temperatures, and format/prepare them for storage.
-
-    Entirely designed for putting into an influxdb database. If you want
-    another database type, well, point it at a different formatting
-    function in the conditional check on tres.
-    """
-    if isinstance(timeoffset, dt.datetime):
-        print("Applying datetime offset: ", timeoffset)
-    elif isinstance(timeoffset, float) or isinstance(timeoffset, int):
-        print("Applying scalar/float offset: ", timeoffset)
-        print("THIS IS AS YET UNHANDLED, OOPS")
-        raise NotImplementedError
-
-    # With what looks like a sample rate of ~10 Hz, 800 samples will give
-    #   me ~80 seconds worth of temperature data.  For a query interval of
-    #   60 seconds, which could vary depending on some picture stuff, this
-    #   will be a pretty good representation of the performance/stability.
-    nsamps = 800
-    endpoint = "/printer/diagnostics/temperature_flow/%d" % (nsamps)
-
-    # This query is a house of cards; if it fails because the printer
-    #   is unreachable, literally everything will implode. So check that
-    #   the return value isn't empty!!!
-    tres = api.queryChecker(printerip, endpoint)
-
-    if tres != {}:
-        # For the Ultimaker 3e, the flow sensor hardware was removed before
-        #   the printer shipped so the following are always 0 or 65535;
-        #   We exclude them from the results because that's annoying.
-        # NOTE: case isn't checked, so they must be *exact* matches!
-        #       Also - "Time" is skipped because we store that differently
-        bklst = ['Time',
-                 'flow_sensor0', 'flow_steps0',
-                 'flow_sensor1', 'flow_steps1']
-
-        allpkts = []
-
-        for i, points in enumerate(tres):
-            # At this point, if the query is successful, tres is a list of
-            #   lists,  the first of which is the labels and the rest are
-            #   lists of values matching those labels.
-            if i == 0:
-                flabs = points
-                # Translate the blacklisted labels to
-                if bklst != []:
-                    gi = [k for k, lab in enumerate(flabs) if lab not in bklst]
-                else:
-                    gi = []
-            else:
-                # Make an influxdb packet, but first do some contortions
-                #   to make the timestamp a real timestamp rather than
-                #   just an offset from boot
-                ts = timeoffset + dt.timedelta(seconds=points[0])
-                # We need to strip out the
-                # Convert it to nanoseconds for influx
-                #   NOTE: It CAN NOT be a float! Must be an Int :(
-                ts = int(float(ts.strftime("%s.%f"))*1e3)
-
-                # Grab the non-blacklisted items and construct our packet
-                #   I can't think of a good way to put this into a set of
-                #   non-confusing list comprehensions so I'm breaking it out
-                fields = {}
-                for idx in gi:
-                    fields.update({flabs[idx]: points[idx]})
-
-                pkt = packetizer.makeInfluxPacket(meas=['temperatures'],
-                                                  ts=ts, fields=fields,
-                                                  tags=None)
-                # Have to do pkt[0] because makeInfluxPacket is still
-                #   annoying and quirky
-                allpkts.append(pkt[0])
-    else:
-        # This happens when the printer query fails
-        allpkts = []
-
-    return allpkts
 
 
 def startCollections(cDict, db=None, loopTime=60):
@@ -154,10 +54,11 @@ def startCollections(cDict, db=None, loopTime=60):
         # Did our status check work?
         if stats != {}:
             # Collect the temperatures
-            tempPkts = tempStats(cDict['printer'].ip, timeoffset=boottimeUTC)
+            tempPkts = printer.tempStats(cDict['printer'].ip,
+                                         timeoffset=boottimeUTC)
 
             # Collect the overall system info
-            sysPkts = systemStats(cDict['printer'].ip)
+            sysPkts = printer.systemStats(cDict['printer'].ip)
 
             if db is not None:
                 if tempPkts != []:
